test(agent-framework): fold returnDirect reproduction into existing test
