+++ spring-ai-alibaba-agent-framework/src/test/java/com/alibaba/cloud/ai/graph/agent/ReactAgentResolverBugReproductionTest.java
@@ -15,16 +15,22 @@
  */
 package com.alibaba.cloud.ai.graph.agent;
 
+import com.alibaba.cloud.ai.graph.agent.hook.returndirect.ReturnDirectModelHook;
 import com.alibaba.cloud.ai.graph.agent.node.AgentLlmNode;
 import com.alibaba.cloud.ai.graph.agent.tools.PoetTool;
 import com.alibaba.cloud.ai.graph.checkpoint.savers.MemorySaver;
+import com.alibaba.cloud.ai.graph.exception.GraphRunnerException;
 import org.junit.jupiter.api.Test;
 import org.springframework.ai.chat.messages.AssistantMessage;
+import org.springframework.ai.chat.messages.Message;
+import org.springframework.ai.chat.messages.ToolResponseMessage;
 import org.springframework.ai.chat.model.ChatModel;
 import org.springframework.ai.chat.model.ChatResponse;
 import org.springframework.ai.chat.model.Generation;
 import org.springframework.ai.chat.prompt.Prompt;
 import org.springframework.ai.tool.ToolCallback;
+import org.springframework.ai.tool.annotation.Tool;
+import org.springframework.ai.tool.annotation.ToolParam;
 import org.springframework.ai.tool.resolution.ToolCallbackResolver;
 import reactor.core.publisher.Flux;
 
@@ -32,6 +38,7 @@ import java.lang.reflect.Field;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -40,6 +47,10 @@ import static org.junit.jupiter.api.Assertions.*;
  * This test verifies that tools provided via resolver are correctly passed to llmNode.
  */
 class ReactAgentResolverBugReproductionTest {
+
+	private static final String DIRECT_TOOL_RESULT = "\"RAW_RESULT=42\"";
+
+	private static final String NATURAL_LANGUAGE_RESULT = "12 加 30 的结果是 42。";
 
 	/**
 	 * Simple ToolCallbackResolver implementation that provides a tool.
@@ -107,6 +118,57 @@ class ReactAgentResolverBugReproductionTest {
 		assertEquals("poem", toolCallbacks.get(0).getToolDefinition().name(), "Tool name should be 'poem'");
 	}
 
+	@Test
+	void testReactAgentReturnDirectShortCircuitsSecondModelCall() throws GraphRunnerException {
+		ScriptedToolLoopChatModel chatModel = new ScriptedToolLoopChatModel();
+
+		ReactAgent agent = ReactAgent.builder()
+				.name("return-direct-default-react-agent")
+				.model(chatModel)
+				.methodTools(new ReturnDirectAddTool())
+				.build();
+
+		AssistantMessage assistantMessage = agent.call("请调用工具计算 12 + 30，并用一句完整中文解释结果");
+
+		assertEquals(DIRECT_TOOL_RESULT, assistantMessage.getText());
+		assertEquals(1, chatModel.callCount.get(), "returnDirect=true should stop before the second model call");
+	}
+
+	@Test
+	void testReactAgentNonReturnDirectStillCallsModelAgain() throws GraphRunnerException {
+		ScriptedToolLoopChatModel chatModel = new ScriptedToolLoopChatModel();
+
+		ReactAgent agent = ReactAgent.builder()
+				.name("return-direct-false-react-agent")
+				.model(chatModel)
+				.methodTools(new NonReturnDirectAddTool())
+				.build();
+
+		AssistantMessage assistantMessage = agent.call("请调用工具计算 12 + 30，并用一句完整中文解释结果");
+
+		assertEquals(NATURAL_LANGUAGE_RESULT, assistantMessage.getText());
+		assertEquals(2, chatModel.callCount.get(), "returnDirect=false should continue into the second model call");
+		assertEquals(DIRECT_TOOL_RESULT, chatModel.lastToolResponseData);
+		assertNotEquals(DIRECT_TOOL_RESULT, assistantMessage.getText());
+	}
+
+	@Test
+	void testExplicitReturnDirectHookStillShortCircuitsSecondModelCall() throws GraphRunnerException {
+		ScriptedToolLoopChatModel chatModel = new ScriptedToolLoopChatModel();
+
+		ReactAgent agent = ReactAgent.builder()
+				.name("return-direct-hooked-react-agent")
+				.model(chatModel)
+				.methodTools(new ReturnDirectAddTool())
+				.hooks(List.of(new ReturnDirectModelHook()))
+				.build();
+
+		AssistantMessage assistantMessage = agent.call("请调用工具计算 12 + 30，并用一句完整中文解释结果");
+
+		assertEquals(DIRECT_TOOL_RESULT, assistantMessage.getText());
+		assertEquals(1, chatModel.callCount.get(), "ReturnDirectModelHook should stop the loop before the second model call");
+	}
+
 	/**
 	 * Helper method to get llmNode from ReactAgent using reflection.
 	 */
@@ -127,5 +189,50 @@ class ReactAgentResolverBugReproductionTest {
 		return toolCallbacks;
 	}
 
-}
+	static class ReturnDirectAddTool {
 
+		@Tool(name = "my_add", description = "实现两个整数相加", returnDirect = true)
+		public String add(@ToolParam(required = true, description = "必须是整数") int first,
+				@ToolParam(required = true, description = "必须是整数") int second) {
+			return "RAW_RESULT=" + (first + second);
+		}
+	}
+
+	static class NonReturnDirectAddTool {
+
+		@Tool(name = "my_add", description = "实现两个整数相加", returnDirect = false)
+		public String add(@ToolParam(required = true, description = "必须是整数") int first,
+				@ToolParam(required = true, description = "必须是整数") int second) {
+			return "RAW_RESULT=" + (first + second);
+		}
+	}
+
+	static final class ScriptedToolLoopChatModel implements ChatModel {
+
+		private final AtomicInteger callCount = new AtomicInteger();
+
+		private volatile String lastToolResponseData;
+
+		@Override
+		public ChatResponse call(Prompt prompt) {
+			int currentCall = callCount.incrementAndGet();
+			if (currentCall == 1) {
+				AssistantMessage.ToolCall toolCall = new AssistantMessage.ToolCall(
+						"call-1", "function", "my_add", "{\"first\":12,\"second\":30}");
+				AssistantMessage assistantMessage = AssistantMessage.builder()
+						.content("")
+						.toolCalls(List.of(toolCall))
+						.build();
+				return new ChatResponse(List.of(new Generation(assistantMessage)));
+			}
+
+			List<Message> instructions = prompt.getInstructions();
+			Message lastMessage = instructions.get(instructions.size() - 1);
+			ToolResponseMessage toolResponseMessage = assertInstanceOf(ToolResponseMessage.class, lastMessage);
+			lastToolResponseData = toolResponseMessage.getResponses().get(0).responseData();
+			System.out.println("[spring-ai-alibaba-repro] second-model-call observed tool response: " + lastToolResponseData);
+			return new ChatResponse(List.of(new Generation(new AssistantMessage(NATURAL_LANGUAGE_RESULT))));
+		}
+	}
+
+}
